flask/requirements.txt
dbus-python
flask
flask-cors
Flask-RESTful
Flask-Jsonpify
flask-inputs
python-vlc 
omxplayer-wrapper